build: update dependencies

+++ python/genvarloader/_dummy.py
@@ -93,7 +93,7 @@ def get_dummy_dataset(spliced: bool = False):
         ]  # target lengths
         - 1  # idx within region
         + 4 * np.arange(4)[:, None]  # adjust by region/contig offset
-    )
+    ).astype(np.int32)
     dummy_genos = SparseGenotypes.from_offsets(
         v_idxs.ravel(),
         (4, 4, 1),
